Added SHA256 code
The firmware is now receiving the correct response from the Avalon Gen2
chips when given a specific input, specified in app.c.

I did some changes to the hardware, thinking that removing the D-type
Flip-flop in the Report signal pre-processing glue logic would improve
things. If the FF is used, the data received is shifted one-bit to the
right (looking at the bits with LSB to the left). We just need to get
the firmware to deal with this. The MSB bit (the last bit to arrive)
should still be seen by the pre-processing circuit even with the FF in.