Partial fix for problem with running out of IRP
Added the "endPoint" field to the IRP,. This will be used later to
access the head of the queue and point it to the next IRP, thereby
discarding the oldest IRP.
If the Raspbian USB driver cooperates, this revision is actually stable.
However, at times the Raspbian USB drivers floods the firmware with
control requests, and the firmware runs out of reply IRPs.
USB de-initialization still does not work. It is commented out for now
app.c.