Found major bug when passing hash input data to chips
For almost a month now I have been trying to get hashing to work
properly. Until now cgminer is still rejecting every hash generated by
the condo miner. The culprit is a bug in which a DWORD array
(sequencedBuffer) is declared in the function APP_Tasks() in app.c. Of
course, declaring it within the function makes the whole array
unretainable between calls to APP_Tasks().