Simply increasing USB_DEVICE_MSGBUF_COUNT to 10 makes USB work much better
Increasing the number of both transmit and receive buffers improves the
stability of USB dramatically. After a while, though, it still
disconnects and reconfigures itself. When this happens, the USB output
buffers are almost full (8 out of 10) of hash messages. It seems that
the cgminer host is not issuing enough IN tokens for the hash results.