Don't enable main interrupt until all peripherals used are initialized
Even when reverting to the previous revision, main initialization fails
to start the main loop. When debugger is aborted, the view does not jump
to the location in the source where the program counter is. If firmware
is allowed to continue, USB remains uninitialized. The only way for USB
to get initialized is to disconnect the USB cable, and reconnect it.

If PICKIT 3 is completely removed and production version is run, USB
works just fine.